requirements.txt
FastAPI
flask
uvicorn
pydantic
pyyaml
rapidfuzz
requests
requests-cache
